Copiable `ProductVariation` and update `Product.stockQuantity` type from `Int?` to `Int64?`

+++ WooCommerce/Classes/ViewRelated/Products/Edit Product/Inventory Settings/ProductInventoryEditableData.swift
@@ -6,7 +6,7 @@ struct ProductInventoryEditableData {
     let sku: String?
     let manageStock: Bool
     let soldIndividually: Bool
-    let stockQuantity: Int?
+    let stockQuantity: Int64?
     let backordersSetting: ProductBackordersSetting?
     let stockStatus: ProductStockStatus?
 }
